// src/generated/models/WriteFileBody.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WriteFileBody = {
    path: string;
    body: string;
    encoding?: string;
    language?: (string | null);
    headers?: (Record<string, any> | null);
};